requirements.txt
Flask==2.3.3
gunicorn==21.2.0
openai==1.13.3
python-dotenv==1.0.1
flask-cors>=3.0.0
openai>=1.0.0
httpx>=0.24.0

